requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
defusedxml==0.8.0rc2
distutils-pytest==0.2.1
Django==5.1.1
django-extensions==3.2.3
django-jalali-date==1.1.3
django-taggit==6.1.0
django-tinymce==4.1.0
environs==11.0.0
Faker==30.0.0
idna==3.10
iniconfig==2.0.0
jalali_core==1.0.0
jdatetime==5.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
oauthlib==3.2.2
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
redis==5.1.0
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==75.1.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
Unidecode==1.3.8
urllib3==2.2.3
Werkzeug==3.0.4
